restaurant_app_backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pip==25.1
PyJWT==2.9.0
setuptools==78.1.1
sqlparse==0.5.2
typing_extensions==4.12.2
wheel==0.45.1
